Proper _UIMenuBuilder swizzle
Now the check for _UIMenuBuilder won't happen every single time an object is initialized.  Not only was this terrible for performance, the original hook is most likely the cause of some odd bugs.  For example, the hook may return an initialized object when it should return NULL and thus altering the control flow.  This change also fixes the endless `[MenuBuilder] Duplicates existing --` warnings when attached to a debugger.